texture changed to be u16s

+++ TileMapRendererExperiments/NewRenderer.h
@@ -7,6 +7,7 @@
 using ArrayTexture2DHandle = u32;
 class Camera2D;
 
+
 struct NewRendererInitialisationInfo {
 	u32 tilemapSizeX;
 	u32 tilemapSizeY;
@@ -21,7 +22,7 @@ public:
 	NewRenderer(const NewRendererInitialisationInfo& info);
 	void DrawChunk(const glm::ivec2& chunkWorldMapOffset, const glm::vec2& pos, const glm::vec2& scale, float rotation, ArrayTexture2DHandle texArray, const Camera2D& camArray) const;
 private:
-	std::unique_ptr<u32[]> m_worldTextureBytes;
+	std::unique_ptr<u16[]> m_worldTextureBytes;
 	u32 m_worldTextureHandle;
 	Shader m_tileShader;
 	glm::ivec2 m_tilemapChunkSize;
